Split off test code that relies on OgContext into a separate base class.
